requirements.txt
discord.py==2.3.2
Flask==2.3.2
gunicorn==21.2.0